--- src/utils/spec.ts
@@ -1,14 +1,0 @@
-import {Spec, NodeSpec, ComponentSpec} from '../defs';
-import {isObject} from './misc';
-
-export function isSpec(x: any): x is Spec {
-    return isObject(x) && x.type != null && x.nodeType == null;
-}
-
-export function isNodeSpec(x: any): x is NodeSpec {
-    return isSpec(x) && typeof x.type === 'string';
-}
-
-export function isComponentSpec(x: any): x is ComponentSpec {
-    return isSpec(x) && typeof x.type === 'function';
-}
